requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.3
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.51.0
django-bootstrap5==21.3
django-environ==0.9.0
idna==3.3
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
requests==2.28.0
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
